test: Use printf-friendly types
In C90, there is no printf type specifier for size_t or ptrdiff_t. Use
printf-friendly types to avoid printf format errors (-Werror=format)
from GCC.

    i586-pc-msdosdjgpp-gcc  -I../src/../include -O3 -DNDEBUG -Wall -Werror -Wextra -std=c89 -pedantic -MD -MT tests/CMakeFiles/test_hashmap.dir/hashmap.c.obj -MF tests/CMakeFiles/test_hashmap.dir/hashmap.c.obj.d -o tests/CMakeFiles/test_hashmap.dir/hashmap.c.obj -c ../tests/hashmap.c
    In file included from ../tests/hashmap.c:10:
    ../tests/hashmap.c: In function 'hashmap_test_get_nonexistent':
    ../tests/test.h:32:21: error: format '%ld' expects argument of type 'long int', but argument 8 has type 'ptrdiff_t' {aka 'int'} [-Werror=format=]
       32 | #define COLOR_START "\x1b[1;%d;%dm"
          |

+++ tests/test.h
@@ -176,21 +176,21 @@ enum
     } while (0)
 
 /* Signed comparison */
-#define TEST_EQ(a, b) TEST_EQ_T(a, b, ptrdiff_t, %ld)
-#define TEST_NE(a, b) TEST_NE_T(a, b, ptrdiff_t, %ld)
-#define TEST_GT(a, b) TEST_GT_T(a, b, ptrdiff_t, %ld)
-#define TEST_LT(a, b) TEST_LT_T(a, b, ptrdiff_t, %ld)
-#define TEST_GE(a, b) TEST_GE_T(a, b, ptrdiff_t, %ld)
-#define TEST_LE(a, b) TEST_LE_T(a, b, ptrdiff_t, %ld)
+#define TEST_EQ(a, b) TEST_EQ_T(a, b, long, %ld)
+#define TEST_NE(a, b) TEST_NE_T(a, b, long, %ld)
+#define TEST_GT(a, b) TEST_GT_T(a, b, long, %ld)
+#define TEST_LT(a, b) TEST_LT_T(a, b, long, %ld)
+#define TEST_GE(a, b) TEST_GE_T(a, b, long, %ld)
+#define TEST_LE(a, b) TEST_LE_T(a, b, long, %ld)
 
 /* Unsigned comparison */
-#define TEST_EQU(a, b) TEST_EQ_T(a, b, size_t, %lu)
-#define TEST_EQX(a, b) TEST_EQ_T(a, b, size_t, 0x%lX)
-#define TEST_NEU(a, b) TEST_NE_T(a, b, size_t, %lu)
-#define TEST_GTU(a, b) TEST_GT_T(a, b, size_t, %lu)
-#define TEST_LTU(a, b) TEST_LT_T(a, b, size_t, %lu)
-#define TEST_GEU(a, b) TEST_GE_T(a, b, size_t, %lu)
-#define TEST_LEU(a, b) TEST_LE_T(a, b, size_t, %lu)
+#define TEST_EQU(a, b) TEST_EQ_T(a, b, unsigned long, %lu)
+#define TEST_EQX(a, b) TEST_EQ_T(a, b, unsigned long, 0x%lX)
+#define TEST_NEU(a, b) TEST_NE_T(a, b, unsigned long, %lu)
+#define TEST_GTU(a, b) TEST_GT_T(a, b, unsigned long, %lu)
+#define TEST_LTU(a, b) TEST_LT_T(a, b, unsigned long, %lu)
+#define TEST_GEU(a, b) TEST_GE_T(a, b, unsigned long, %lu)
+#define TEST_LEU(a, b) TEST_LE_T(a, b, unsigned long, %lu)
 
 /* Pointer comparison */
 #define TEST_EQP(a, b) TEST_EQ_T(a, b, void *, %p)
